fix: Fix types.

+++ activebits.go
@@ -1,8 +1,8 @@
 package piscine
 
-func ActiveBits(n int) uint {
+func ActiveBits(n int) int {
 	if 0 <= n && n < 2 {
-		return uint(n)
+		return int(uint(n))
 	}
-	return (uint(n) % 2) + ActiveBits(n/2)
+	return int((uint(n) % 2) + uint(ActiveBits(n/2)))
 }
